Clean-up properly on container shutdown
This code fixes the previous "wait" approach by starting the sleep process in the background, so the wait actually gets to wait on something.

The problem before was that the sleep process was the "foreground" process (not bash), so kill signals from docker went to the sleep process instead of bash, which in turn meant that the "finish" trap was never triggered and run. With this fix, the wait (aka bash) is in the foreground and when docker shuts down the container, bash will receive the kill signal, stop waiting and run the finish function.

I also changed the sleep from 1 day to infinity, so no loop is necessary.